pyproject.toml
[project]
name = "lab0-titanic-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "joblib>=1.5.3",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "ruff>=0.15.11",
]
[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint]
select = ["E", "F", "W"]
